# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ResolveAI resolution copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: getCases
      tags: [cases]
      summary: List cases for agents
      parameters:
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Priority"
        - $ref: "#/components/parameters/Category"
        - name: escalated
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a case from a customer issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseInput"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
        "400":
          $ref: "#/components/responses/Error"
...
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case and its full history
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /cases/{id}/messages:
    post:
      operationId: addCaseMessage
      tags: [cases]
      summary: Add a message to a case
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /cases/{id}/attachments:
    post:
      operationId: uploadAttachment
      tags: [cases]
      summary: Upload evidence to a case
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AttachmentUploadInput"
      responses:
        "201":
          description: Attachment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
  /attachments/{id}/url:
    get:
      operationId: getAttachmentUrl
      tags: [cases]
      summary: Get a signed URL for an attachment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signed URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentUrlResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /cases/{id}/generate-reply:
    post:
      operationId: generateReply
      tags: [cases]
      summary: Generate an agent reply draft
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "200":
          description: Draft reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyDraft"
  /cases/{id}/send-reply:
    post:
      operationId: sendReply
      tags: [cases]
      summary: Send an agent reply
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /cases/{id}/status:
    patch:
      operationId: updateCaseStatus
      tags: [cases]
      summary: Change case status
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusInput"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}/override:
    patch:
      operationId: overrideCase
      tags: [cases]
      summary: Override an AI field
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverrideInput"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}/assign:
    post:
      operationId: assignCase
      tags: [cases]
      summary: Assign a case to an agent
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignInput"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Order:
      type: object
      properties:
        id: { type: integer }
        order_code: { type: string }
        user_id: { type: integer }
        product_name: { type: string }
        price: { type: number }
        order_date: { type: string }
        delivery_date: { type: string }
        status: { type: string }
      required: [id, order_code, user_id, product_name, price, order_date, delivery_date, status]
    Case:
      type: object
      properties:
        id: { type: integer }
        case_code: { type: string }
        user_id: { type: integer }
        order_id: { type: ["integer", "null"] }
        status: { type: string }
        category: { type: string }
        priority: { type: string }
        sentiment: { type: string }
        customer_intent: { type: string }
        summary: { type: string }
        recommended_action: { type: string }
        next_step: { type: string }
        escalation_required: { type: boolean }
        escalation_reason: { type: ["string", "null"] }
        missing_information:
          type: array
          items: { type: string }
        resolution_plan:
          type: array
          items: { type: string }
        assigned_agent_id: { type: ["integer", "null"] }
        ai_failed: { type: boolean }
        created_at: { type: string }
        updated_at: { type: string }
        due_at: { type: string }
        customer_name: { type: ["string", "null"] }
        order: { anyOf: [{ $ref: "#/components/schemas/Order" }, { type: "null" }] }
      required:
        [id, case_code, user_id, order_id, status, category, priority, sentiment, customer_intent, summary, recommended_action, next_step, escalation_required, missing_information, resolution_plan, ai_failed, created_at, updated_at, due_at]
    CaseInput:
      type: object
      properties:
        message: { type: string, minLength: 10 }
        order_id: { type: ["integer", "null"] }
      required: [message]
    CaseDetail:
      allOf:
        - $ref: "#/components/schemas/Case"
        - type: object
          properties:
            timeline:
              type: array
              items: { $ref: "#/components/schemas/CaseEvent" }
            messages:
              type: array
              items: { $ref: "#/components/schemas/Message" }
            attachments:
              type: array
              items: { $ref: "#/components/schemas/Attachment" }
          required: [timeline, messages, attachments]
    CaseEvent:
      type: object
      properties:
        id: { type: integer }
        case_id: { type: integer }
        event_type: { type: string }
        description: { type: string }
        actor_type: { type: string }
        created_at: { type: string }
        metadata: { type: object, additionalProperties: true }
      required: [id, case_id, event_type, description, actor_type, created_at]
    Message:
      type: object
      properties:
        id: { type: integer }
        case_id: { type: integer }
        sender_type: { type: string, enum: [customer, agent] }
        body: { type: string }
        created_at: { type: string }
      required: [id, case_id, sender_type, body, created_at]
    MessageInput:
      type: object
      properties:
        body: { type: string, minLength: 1 }
      required: [body]
    Attachment:
      type: object
      properties:
        id: { type: integer }
        case_id: { type: integer }
        file_name: { type: string }
        file_path: { type: string }
        storage_path: { type: string }
        mime_type: { type: string }
        size_bytes: { type: integer }
        uploaded_at: { type: string }
      required: [id, case_id, file_name, file_path, uploaded_at]
    AttachmentUrlResponse:
      type: object
      properties:
        url: { type: string }
        expires_in: { type: integer }
      required: [url, expires_in]
    AttachmentUploadInput:
      type: object
      properties:
        file_name: { type: string, minLength: 1 }
        file_data: { type: string, minLength: 1 }
      required: [file_name, file_data]
    ReplyInput:
      type: object
      properties:
        body: { type: string, minLength: 1 }
        tone: { type: string, enum: [warm, concise, professional] }
      required: [body]
    ReplyDraft:
      type: object
      properties:
        draft: { type: string }
      required: [draft]
    StatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [New, Investigating, Awaiting Customer, Approved, Resolved, Rejected]
      required: [status]
    OverrideInput:
      type: object
      properties:
        field: { type: string }
        value: { type: string }
      required: [field, value]
    AssignInput:
      type: object
      properties:
        agent_id: { type: integer }
      required: [agent_id]
...
  parameters:
    CaseId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    Status:
      name: status
      in: query
      schema:
        type: string
        enum: [New, Investigating, Awaiting Customer, Approved, Resolved, Rejected]
    Priority:
      name: priority
      in: query
      schema:
        type: string
        enum: [Low, Medium, High, Critical]
    Category:
      name: category
      in: query
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"